Setting the status to passing for rebooting rds instances
Instances do not become unavailable in the seconds the status changes to rebooting,
so de-registering the service from consul can actually be more harm that worth.

Additionally, if the RDS instance is configured with Multi-AZ failover, rebooting the
instance should actaully never mean a change in the passing status of the service.